+++ UI/Widgets/Button.lua
@@ -1,0 +1,13 @@
+local _, addon = ...
+
+local name = "Button"
+local Object = {}
+
+function Object:New(parent, name)
+   local b = addon.UI.CreateFrame("Button", name, parent, "UIPanelButtonTemplate")
+	b:SetText("")
+	b:SetSize(100,25)
+	return b
+end
+
+addon.UI:RegisterElement(Object, name)