+++ changelog.md
@@ -5,9 +5,26 @@
 * A warning is printed if a candidate trades an awarded item to the wrong person.
 
 
+* **Personal Loot**
+* The council can now see all items looted whether tradeable or not.
+* New buttons will appear for any items that can't be added to the session under the voting frame.
+* Usage options have been updated to only include PL options.
+
+
+* **Allow Keeping**
+* The Group Leader can now choose whether candidates can keep their items.
+* If enabled (disabled by default) candidates will see a popup asking if the want to keep the loot whenever they loot a tradeable item.
+
+
 ###### Bugfixes
 * *Removed usage popup for non group leaders.*
+* *Usage popup should be more reliable (#350).*
 
+
+###### Dev
+* Two new comm messages (`not_tradeable` and `rejected_trade`).
+* Added UI section. This is the beginning of a consolidation of UI elements - everything will use this format soon ish.
+* Backwards compatibility isn't broken yet, but will be at some point.
 
 ### v2.8.2
 ---